Added Team.h framework
Added team.h and included it in main. It says I changed the data.csv
file but I just looked at it, even if I did change it we should update
it the day before the presentation. Also included a file to decode my
variable names if they are confusing.